pyproject.toml
[project]
name = "redstone-rom-programmer"
version = "0.1.0"
description = "Redstone Rom Programmer"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "intelhex>=2.3.0",
    "mcschematic>=11.4.4",
    "pyyaml>=6.0.3",
]
